Update test structure
