+++ src/__tests__/create.test.js
@@ -3,11 +3,11 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import Login from './login'
+import CreateQuestion from '../pages/questions/create'
 
 describe('Snapshot-Testing', () => {
-  it('Works', () => {
-    const component = renderer.create(<Login />)
+  it.skip('Works', () => {
+    const component = renderer.create(<CreateQuestion />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
