+++ src/__tests__/join.test.js
@@ -3,11 +3,11 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import CreateQuestion from './create'
+import Join from '../pages/join'
 
 describe('Snapshot-Testing', () => {
-  it.skip('Works', () => {
-    const component = renderer.create(<CreateQuestion />)
+  it('Works', () => {
+    const component = renderer.create(<Join url={{ query: { shortname: 'rsc' } }} />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
