+++ src/__tests__/login.test.js
@@ -3,11 +3,11 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import Registration from './registration'
+import Login from '../pages/user/login'
 
 describe('Snapshot-Testing', () => {
-  it.skip('Works', () => {
-    const component = renderer.create(<Registration />)
+  it('Works', () => {
+    const component = renderer.create(<Login />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
