+++ src/__tests__/registration.test.js
@@ -3,11 +3,11 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import Index from './index'
+import Registration from '../pages/user/registration'
 
 describe('Snapshot-Testing', () => {
   it.skip('Works', () => {
-    const component = renderer.create(<Index />)
+    const component = renderer.create(<Registration />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
