+++ src/__tests__/running.test.js
@@ -3,10 +3,10 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import Running from './running'
+import Running from '../pages/sessions/running'
 
 describe('Snapshot-Testing', () => {
-  it.skip('Works', () => {
+  it('Works', () => {
     const component = renderer.create(<Running />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
